Move @Auxiliatrix TokenizerTest with other Tokenizer tests
