+++ src/test/java/scilib/utilities/TokenizerTest.java
@@ -8,7 +8,30 @@ import java.util.Iterator;
 import org.junit.Ignore;
 import org.junit.Test;
 
+/**
+ * A collection of tests for the Tokenizer utility.
+ * @author Auxiliatrix
+ * @author rlee287
+ */
 public class TokenizerTest {
+
+    @Test
+    public void tokenizerParse() {
+        Tokenizer t = new Tokenizer();
+        ArrayList<String> expected = new ArrayList<String>();
+        expected.add("A");
+        expected.add("B");
+        expected.add("C");
+        expected.add("D");
+        expected.add("E");
+        expected.add("F");
+        expected.add("G");
+        expected.add("H");
+        expected.add("I");
+        expected.add("J");
+        expected.add("K");
+        assertEquals(expected, t.parse("A B \"C\" \"D\" E\"F\" G\"H\" I\"J\"K"));
+    }
 
 	@Test
 	public void testParse() {
